requirements.txt
Flask==3.1.1
Flask-SocketIO==5.5.1
requests==2.32.4
SQLAlchemy==2.0.42
solana==0.30.4
solders==0.6.0
waitress